Code functioning
This commit contains working code to create a password based on some user defined options.

Alert box created to ask user to enter number of characters.
Modal popup created with checkboxes to allow the user to select what type of characters are to be in the password.
Alerts created to inform user when the code needs to quit because an innapproprate selection was received.
On event code was created in javascript:
- on button click event to start the code running
- on modal close to collect the values of the check boxes the user has entered
- on modal closed to check if an appropriate selection was made and continue the code
IF time (and desire) permits I will change the standard alerts to modal popups so they have the same styling.  But I may just do this when I re-use modal dialogs in future.
Some minor changes to the style sheet were implemented to ensure consistency.